pyproject.toml
[project]
name = "decensorer"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "flask>=3.1.2",
    "pillow>=11.3.0",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
